Emprolijo Ganado-Huevo-Pollito y agrego un par desde DDBB

+++ src/main/java/com/accenture/granja/model/Producto.java
@@ -9,14 +9,14 @@ import java.time.LocalDate;
 
 public abstract class Producto {
 	LocalDate now = LocalDate.now();
-	private int id;
+	private long id;
 	private String animal;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
@@ -47,7 +47,7 @@ public abstract class Producto {
 	private LocalDate fechaExpiracion;
 	private BigDecimal precio;
 
-	public Producto(int id, String animal, LocalDate fechaExpiracion, BigDecimal precio) {
+	public Producto(long id, String animal, LocalDate fechaExpiracion, BigDecimal precio) {
 		this.id = id;
 		this.fechaExpiracion = fechaExpiracion;
 		this.precio = precio;
